Reorganize ECS components and systems
