+++ src/server/ecs/systems/server/updateTransforms.ts
@@ -1,6 +1,6 @@
 import { World } from "@rbxts/matter";
 import { Workspace } from "@rbxts/services";
-import { Model, Transform } from "shared/components";
+import { Model, Transform } from "shared/ecs/components";
 import removeMissingModels from "./removeMissingModels";
 
 function updateTransforms(world: World): void {
