+++ src/shared/ecs/components.spec.ts
