+++ src/shared/ecs/components/defaults.ts
