+++ src/shared/ecs/components/index.ts
