+++ src/shared/ecs/components/types.d.ts
